Workaround for possible PLCrashReporter crashes
In some very rare cases, PLCrashReporter may have issues parsing a crash report and crash itself. This commit will make sure, that this happens only once at max for a single crash, and the problematic file will be deleted on the next startup.
In general, that kind of crash can be fixed, by adding -all_load to the linker flags.