Add AutoSubmitDeviceUDID property
This property will tell Quincy if the device UDID shall be used as user ID and send as part of the crash report. It is basically a shortcut instead of implementing the crashReportUserID delegate and adding the UDID in there. Default is NO.
This should only be used in beta versions!