bluetooth evidence source ready for testing
A lot of changes in an effort to get the bluetooth evidence source more
stable.  This is ready for testing but there is one outstanding, known
issue.  If the bluetooth evidence source is disabled ControlPlane will
continue to respond to inquiryDidComplete events which is evident in
console logs.  The only way to resolve this is to restart ControlPlane.
I don't feel like this is too much of a show stopper as most people will
enable it and leave it alone.    The bluetooth evidence source should
respond to connection notifications and be able to scan for nearby
bluetooth devices with discoverable enabled.  It should be able to
tolerate bluetooth being disabled as well as a sleep/wake cycle.